+++ skymap_scanner/recos/millipede_wilks.py
@@ -29,7 +29,7 @@ from icecube.icetray import I3Frame
 from .. import config as cfg
 from ..utils.pixel_classes import RecoPixelVariation
 from . import RecoInterface, VertexGenerator
-from .common.pulse_proc import mask_deepcore
+from .common.pulse_proc import mask_deepcore, late_pulse_cleaning
 
 class MillipedeWilks(RecoInterface):
     """Reco logic for millipede."""
@@ -177,56 +177,7 @@ class MillipedeWilks(RecoInterface):
 
         ##################
 
-        def _weighted_quantile_arg(values, weights, q=0.5):
-            indices = numpy.argsort(values)
-            sorted_indices = numpy.arange(len(values))[indices]
-            medianidx = (weights[indices].cumsum()/weights[indices].sum()).searchsorted(q)
-            if (0 <= medianidx) and (medianidx < len(values)):
-                return sorted_indices[medianidx]
-            else:
-                return numpy.nan
-
-        def weighted_quantile(values, weights, q=0.5):
-            if len(values) != len(weights):
-                raise ValueError("shape of `values` and `weights` don't match!")
-            index = _weighted_quantile_arg(values, weights, q=q)
-            if not numpy.isnan(index):
-                return values[index]
-            else:
-                return numpy.nan
-
-        def weighted_median(values, weights):
-            return weighted_quantile(values, weights, q=0.5)
-
-        def LatePulseCleaning(frame, Pulses, Residual=1.5e3*I3Units.ns):
-            pulses = dataclasses.I3RecoPulseSeriesMap.from_frame(frame, Pulses)
-            mask = dataclasses.I3RecoPulseSeriesMapMask(frame, Pulses)
-            counter, charge = 0, 0
-            qtot = 0
-            times = dataclasses.I3TimeWindowSeriesMap()
-            for omkey, ps in pulses.items():
-                if len(ps) < 2:
-                    if len(ps) == 1:
-                        qtot += ps[0].charge
-                    continue
-                ts = numpy.asarray([p.time for p in ps])
-                cs = numpy.asarray([p.charge for p in ps])
-                median = weighted_median(ts, cs)
-                qtot += cs.sum()
-                for p in ps:
-                    if p.time >= (median+Residual):
-                        if omkey not in times:
-                            ts = dataclasses.I3TimeWindowSeries()
-                            ts.append(dataclasses.I3TimeWindow(median+Residual, numpy.inf)) # this defines the **excluded** time window
-                            times[omkey] = ts
-                        mask.set(omkey, p, False)
-                        counter += 1
-                        charge += p.charge
-            frame[cls.pulsesName_cleaned] = mask
-            frame[cls.pulsesName_cleaned+"TimeWindows"] = times
-            frame[cls.pulsesName_cleaned+"TimeRange"] = copy.deepcopy(frame[cls.pulsesName_orig+"TimeRange"])
-
-        tray.AddModule(LatePulseCleaning, "LatePulseCleaning",
+        tray.AddModule(late_pulse_cleaning, "LatePulseCleaning",
                        Pulses=cls.pulsesName,
                        )
         return ExcludedDOMs + [cls.pulsesName_cleaned+'TimeWindows']
